完成代码美化工作的最终提交：工具模块优化
- 完成了工具类模块(06_utilities.c)的最终优化
- 统一了整个代码库的命名规范和代码风格
- 确保所有模块的代码风格一致性
- 保持了原有代码逻辑完整性，仅进行命名和注释优化

+++ TaleWorlds.Native/src/01_initialization.c
@@ -19687,10 +19687,10 @@ void ProcessSystemThreeParameterBuffer(long long MainParameter,long long Auxilia
   uint8_t* SystemBufferPointer;
   uint32_t SystemBufferLength;
   uint8_t SystemDataBuffer [1032];
-  ulong long SystemChecksumValue;
+  unsigned long long SystemChecksumValue;
   
   SystemSecurityFlag = 0xfffffffffffffffe;
-  SystemChecksumValue = SystemEncryptionKeyTemplate ^ (ulong long)StackSecurityBuffer;
+  SystemChecksumValue = SystemEncryptionKeyTemplate ^ (unsigned long long)StackSecurityBuffer;
   SystemMemoryReference = &SystemMemoryTemplateF;
   SystemBufferPointer = SystemDataBuffer;
   SystemBufferLength = 0;
